Refactor for bitshifting and wiring fix for ln
Refactored the bitshifting VI to make logic more clear and make handshake outputs only depend on current state.

Fixed bad wiring in ln.vi where a select for an FF was wired to output valid instead of input valid.

All that is left is implementing multiple accesses which will be straighforward.