Added a new version of cache gen and updated code so it will compile
The new cache gen makes my life easier as it shortens the generation process to 2 steps (run and select folder) and all files are put in their proper location.

Updated cache gen code for includes to be correct and fixed syntax errors (missing '&' in C and missing f for f strings in python)

Updated the state format to use macros for logging.